move all functions related to data loading and checking to diffractometer.py
