+++ bcdi/preprocessing/preprocessing_utils.py
@@ -14,13 +14,13 @@ except ModuleNotFoundError:
     pass
 
 import matplotlib.pyplot as plt
-from numbers import Real, Integral
+from numbers import Real
 import numpy as np
 from scipy.ndimage.measurements import center_of_mass
 from scipy.interpolate import RegularGridInterpolator
-import tkinter as tk
-from tkinter import filedialog
 import xrayutilities as xu
+
+from ..experiment import diffractometer as diff
 from ..graph import graph_utils as gu
 from ..utils import image_registration as reg
 from ..utils import utilities as util
@@ -798,183 +798,6 @@ def center_fft(
     return data, mask, pad_width, q_values, frames_logical
 
 
-def check_empty_frames(data, mask=None, monitor=None, frames_logical=None):
-    """
-    Check if there is intensity for all frames.
-
-    In case of beam dump, some frames may be empty. The data and optional mask will be
-    cropped to remove those empty frames.
-
-    :param data: a numpy 3D array
-    :param mask: a numpy 3D array of 0 (pixel not masked) and 1 (masked pixel),
-     same shape as data
-    :param monitor: a numpy 1D array of shape equal to data.shape[0]
-    :param frames_logical: 1D array of length equal to the number of measured frames.
-     In case of cropping the length of the stack of frames changes. A frame whose
-     index is set to 1 means that it is used, 0 means not used.
-    :return:
-     - cropped data as a numpy 3D array
-     - cropped mask as a numpy 3D array
-     - cropped monitor as a numpy 1D array
-     - updated frames_logical
-
-    """
-    valid.valid_ndarray(arrays=data, ndim=3)
-    if mask is not None:
-        valid.valid_ndarray(arrays=mask, shape=data.shape)
-    if monitor is not None:
-        if not isinstance(monitor, np.ndarray):
-            raise TypeError("monitor should be a numpy array")
-        if monitor.ndim != 1 or len(monitor) != data.shape[0]:
-            raise ValueError("monitor be a 1D array of length data.shae[0]")
-
-    if frames_logical is None:
-        frames_logical = np.ones(data.shape[0])
-    valid.valid_1d_array(
-        frames_logical,
-        allowed_types=Integral,
-        allow_none=False,
-        allowed_values=(0, 1),
-        name="frames_logical",
-    )
-
-    # check if there are empty frames
-    is_intensity = np.zeros(data.shape[0])
-    is_intensity[np.argwhere(data.sum(axis=(1, 2)))] = 1
-    if is_intensity.sum() != data.shape[0]:
-        print("\nEmpty frame detected, cropping the data\n")
-
-    # update frames_logical
-    frames_logical = np.multiply(frames_logical, is_intensity)
-
-    # remove empty frames from the data and update the mask and the monitor
-    data = data[np.nonzero(frames_logical)]
-    mask = mask[np.nonzero(frames_logical)]
-    monitor = monitor[np.nonzero(frames_logical)]
-    return data, mask, monitor, frames_logical
-
-
-def check_pixels(data, mask, debugging=False):
-    """
-    Check for hot pixels in the data using the mean value and the variance.
-
-    :param data: 3D diffraction data
-    :param mask: 2D or 3D mask. Mask will summed along the first axis if a 3D array.
-    :param debugging: set to True to see plots
-    :type debugging: bool
-    :return: the filtered 3D data and the updated 2D mask.
-    """
-    valid.valid_ndarray(arrays=data, ndim=3)
-    valid.valid_ndarray(arrays=mask, ndim=(2, 3))
-    nbz, nby, nbx = data.shape
-
-    if mask.ndim == 3:  # 3D array
-        print("Mask is a 3D array, summing it along axis 0")
-        mask = mask.sum(axis=0)
-        mask[np.nonzero(mask)] = 1
-    valid.valid_ndarray(arrays=mask, shape=(nby, nbx))
-
-    print(
-        "\ncheck_pixels(): number of masked pixels due to detector gaps ="
-        f" {int(mask.sum())} on a total of {nbx*nby}"
-    )
-    meandata = data.mean(axis=0)  # 2D
-    vardata = 1 / data.var(axis=0)  # 2D
-    var_mean = vardata[vardata != np.inf].mean()
-    vardata[meandata == 0] = var_mean
-    # pixels were data=0 (i.e. 1/variance=inf) are set to the mean of  1/var:
-    # we do not want to mask pixels where there was no intensity during the scan
-
-    if debugging:
-        gu.combined_plots(
-            tuple_array=(mask, meandata, vardata),
-            tuple_sum_frames=False,
-            tuple_sum_axis=0,
-            tuple_width_v=None,
-            tuple_width_h=None,
-            tuple_colorbar=True,
-            tuple_vmin=0,
-            tuple_vmax=(1, 1, np.nan),
-            tuple_scale=("linear", "linear", "linear"),
-            tuple_title=(
-                "Input mask",
-                "check_pixels()\nmean(data) before masking",
-                "check_pixels()\n1/var(data) before masking",
-            ),
-            reciprocal_space=True,
-            position=(131, 132, 133),
-        )
-
-    # calculate the mean and variance of a single photon event along the rocking curve
-    min_count = 0.99  # pixels with only 1 photon count along the rocking curve,
-    # use the value 0.99 to be inclusive
-    mean_singlephoton = min_count / nbz
-    var_singlephoton = (
-        ((nbz - 1) * mean_singlephoton ** 2 + (min_count - mean_singlephoton) ** 2)
-        * 1
-        / nbz
-    )
-    print(
-        "check_pixels(): var_mean={:.2f}, 1/var_threshold={:.2f}".format(
-            var_mean, 1 / var_singlephoton
-        )
-    )
-
-    # mask hotpixels with zero variance
-    temp_mask = np.zeros((nby, nbx))
-    temp_mask[vardata == np.inf] = 1
-    # this includes only hotpixels since zero intensity pixels were set to var_mean
-    mask[np.nonzero(temp_mask)] = 1  # update the mask with zero variance hotpixels
-    vardata[vardata == np.inf] = 0  # update the array
-    print(
-        "check_pixels(): number of zero variance hotpixels = {:d}".format(
-            int(temp_mask.sum())
-        )
-    )
-
-    # filter out pixels which have a variance smaller that the threshold
-    # (note that  vardata = 1/data.var())
-    indices_badpixels = np.nonzero(vardata > 1 / var_singlephoton)
-    mask[indices_badpixels] = 1  # mask is 2D
-    print(
-        "check_pixels(): number of pixels with too low variance = {:d}\n".format(
-            indices_badpixels[0].shape[0]
-        )
-    )
-
-    # update the data array
-    indices_badpixels = np.nonzero(mask)  # update indices
-    for index in range(nbz):
-        tempdata = data[index, :, :]
-        tempdata[
-            indices_badpixels
-        ] = 0  # numpy array is mutable hence data will be modified
-
-    if debugging:
-        meandata = data.mean(axis=0)
-        vardata = 1 / data.var(axis=0)
-        vardata[meandata == 0] = var_mean  # 0 intensity pixels, not masked
-        gu.combined_plots(
-            tuple_array=(mask, meandata, vardata),
-            tuple_sum_frames=False,
-            tuple_sum_axis=0,
-            tuple_width_v=None,
-            tuple_width_h=None,
-            tuple_colorbar=True,
-            tuple_vmin=0,
-            tuple_vmax=(1, 1, np.nan),
-            tuple_scale="linear",
-            tuple_title=(
-                "Output mask",
-                "check_pixels()\nmean(data) after masking",
-                "check_pixels()\n1/var(data) after masking",
-            ),
-            reciprocal_space=True,
-            position=(131, 132, 133),
-        )
-    return data, mask
-
-
 def find_bragg(data, peak_method):
     """
     Find the Bragg peak position in data based on the centering method.
@@ -1465,7 +1288,7 @@ def load_bcdi_data(
         name="preprocessing_utils.load_bcdi_data",
     )
 
-    rawdata, rawmask, monitor, frames_logical = load_data(
+    rawdata, rawmask, monitor, frames_logical = setup.diffractometer.load_check_dataset(
         logfile=logfile,
         scan_number=scan_number,
         detector=detector,
@@ -1545,278 +1368,6 @@ def load_bcdi_data(
     return rawdata, rawmask, frames_logical, monitor
 
 
-def load_data(
-    logfile,
-    scan_number,
-    detector,
-    setup,
-    flatfield=None,
-    hotpixels=None,
-    background=None,
-    normalize="skip",
-    bin_during_loading=False,
-    debugging=False,
-):
-    """
-    Load data, apply filters and concatenate it for phasing.
-
-    :param logfile: the logfile created in Setup.create_logfile()
-    :param scan_number: the scan number to load
-    :param detector: an instance of the class Detector
-    :param setup: an instance of the class Setup
-    :param flatfield: the 2D flatfield array
-    :param hotpixels: the 2D hotpixels array. 1 for a hotpixel, 0 for normal pixels.
-    :param background: the 2D background array to subtract to the data
-    :param normalize: 'monitor' to return the default monitor values, 'sum_roi' to
-     return a monitor based on the integrated intensity in the region of interest
-     defined by detector.sum_roi, 'skip' to do nothing
-    :param bin_during_loading: if True, the data will be binned in the detector frame
-     while loading. It saves a lot of memory for large detectors.
-    :param debugging: set to True to see plots
-    :return:
-
-     - the 3D data array in the detector frame and the 3D mask array
-     - the monitor values for normalization
-     - frames_logical: 1D array of length equal to the number of measured frames.
-       In case of cropping the length of the stack of frames changes. A frame whose
-       index is set to 1 means that it is used, 0 means not used.
-
-    """
-    print(
-        "User-defined ROI size (VxH):",
-        detector.roi[1] - detector.roi[0],
-        detector.roi[3] - detector.roi[2],
-    )
-    print(
-        "Detector physical size without binning (VxH):",
-        detector.nb_pixel_y,
-        detector.nb_pixel_x,
-    )
-    print(
-        "Detector size with binning (VxH):",
-        detector.nb_pixel_y // detector.binning[1],
-        detector.nb_pixel_x // detector.binning[2],
-    )
-
-    if setup.filtered_data:
-        data, mask3d, monitor, frames_logical = load_filtered_data(detector=detector)
-
-    else:
-        data, mask2d, monitor, frames_logical = setup.diffractometer.load_data(
-            logfile=logfile,
-            setup=setup,
-            scan_number=scan_number,
-            detector=detector,
-            flatfield=flatfield,
-            hotpixels=hotpixels,
-            background=background,
-            normalize=normalize,
-            bin_during_loading=bin_during_loading,
-            debugging=debugging,
-        )
-
-        # bin the 2D mask if necessary
-        if bin_during_loading:
-            mask2d = util.bin_data(
-                mask2d, (detector.binning[1], detector.binning[2]), debugging=False
-            )
-            mask2d[np.nonzero(mask2d)] = 1
-
-        # check for abnormally behaving pixels
-        data, mask2d = check_pixels(data=data, mask=mask2d, debugging=debugging)
-        mask3d = np.repeat(mask2d[np.newaxis, :, :], data.shape[0], axis=0)
-        mask3d[np.isnan(data)] = 1
-        data[np.isnan(data)] = 0
-
-        # check for empty frames (no beam)
-        data, mask3d, monitor, frames_logical = check_empty_frames(
-            data=data, mask=mask3d, monitor=monitor, frames_logical=frames_logical
-        )
-
-        # intensity normalization
-        if normalize == "skip":
-            print("Skip intensity normalization")
-        else:
-            print("Intensity normalization using " + normalize)
-            data, monitor = util.normalize_dataset(
-                array=data,
-                monitor=monitor,
-                norm_to_min=True,
-                savedir=detector.savedir,
-                debugging=debugging,
-            )
-
-    return data, mask3d, monitor, frames_logical.astype(int)
-
-
-def load_filtered_data(detector):
-    """
-    Load a filtered dataset and the corresponding mask.
-
-    :param detector: an instance of the class Detector
-    :return: the data and the mask array
-    """
-    root = tk.Tk()
-    root.withdraw()
-
-    file_path = filedialog.askopenfilename(
-        initialdir=detector.datadir,
-        title="Select data file",
-        filetypes=[("NPZ", "*.npz")],
-    )
-    data = np.load(file_path)
-    npz_key = data.files
-    data = data[npz_key[0]]
-    file_path = filedialog.askopenfilename(
-        initialdir=detector.datadir,
-        title="Select mask file",
-        filetypes=[("NPZ", "*.npz")],
-    )
-    mask = np.load(file_path)
-    npz_key = mask.files
-    mask = mask[npz_key[0]]
-
-    monitor = np.ones(data.shape[0])
-    frames_logical = np.ones(data.shape[0])
-
-    return data, mask, monitor, frames_logical
-
-
-def mean_filter(
-    data,
-    nb_neighbours,
-    mask=None,
-    target_val=0,
-    extent=1,
-    min_count=3,
-    interpolate="mask_isolated",
-    debugging=False,
-):
-    """
-    Mask or apply a mean filter to data.
-
-    The procedure is applied only if the empty pixel is surrounded by nb_neighbours or
-    more pixels with at least min_count intensity per pixel.
-
-    :param data: 2D or 3D array to be filtered
-    :param nb_neighbours: minimum number of non-zero neighboring pixels for median
-     filtering
-    :param mask: mask array of the same shape as data
-    :param target_val: value where to interpolate, allowed values are int>=0 or np.nan
-    :param extent: in pixels, extent of the averaging window from the reference pixel
-     (extent=1, 2, 3 ... corresponds to window width=3, 5, 7 ... )
-    :param min_count: minimum intensity (inclusive) in the neighboring pixels to
-     interpolate, the pixel will be masked otherwise.
-    :param interpolate: based on 'nb_neighbours', if 'mask_isolated' will mask
-     isolated pixels, if 'interp_isolated' will interpolate isolated pixels
-    :param debugging: set to True to see plots
-    :type debugging: bool
-    :return: updated data and mask, number of pixels treated
-    """
-    valid.valid_ndarray(arrays=data, ndim=(2, 3))
-    # check some mparameters
-    if mask is None:
-        mask = np.zeros(data.shape)
-    valid.valid_ndarray(arrays=mask, shape=data.shape)
-
-    if not np.isnan(target_val) and not isinstance(target_val, int):
-        raise ValueError(
-            "target_val should be nan or an integer, cannot assess float equality"
-        )
-
-    valid.valid_item(
-        nb_neighbours, allowed_types=int, min_excluded=0, name="mean_filter"
-    )
-    valid.valid_item(extent, allowed_types=int, min_excluded=0, name="mean_filter")
-    valid.valid_item(min_count, allowed_types=int, min_included=0, name="mean_filter")
-
-    if interpolate not in {"mask_isolated", "interp_isolated"}:
-        raise ValueError(
-            f"invalid value '{interpolate}' for interpolate,"
-            f" allowed are 'mask_isolated' and 'interp_isolated'"
-        )
-    if not isinstance(debugging, bool):
-        raise TypeError(f"debugging should be a boolean, got {type(debugging)}")
-
-    # find all voxels to be processed
-    if target_val is np.nan:
-        target_pixels = np.argwhere(np.isnan(data))
-    else:
-        target_pixels = np.argwhere(data == target_val)
-    nb_pixels = 0
-
-    if debugging:
-        gu.combined_plots(
-            tuple_array=(data, mask),
-            tuple_sum_frames=(False, False),
-            tuple_sum_axis=(0, 0),
-            tuple_width_v=(None, None),
-            tuple_width_h=(None, None),
-            tuple_colorbar=(True, True),
-            tuple_vmin=(-1, 0),
-            tuple_vmax=(np.nan, 1),
-            tuple_scale=("log", "linear"),
-            tuple_title=("Data before filtering", "Mask before filtering"),
-            reciprocal_space=True,
-        )
-
-    if data.ndim == 2:
-        for indx in range(target_pixels.shape[0]):
-            pixrow = target_pixels[indx, 0]
-            pixcol = target_pixels[indx, 1]
-            temp = data[
-                pixrow - extent : pixrow + extent + 1,
-                pixcol - extent : pixcol + extent + 1,
-            ]
-            temp = temp[np.logical_and(~np.isnan(temp), temp != target_val)]
-            if (
-                temp.size >= nb_neighbours and (temp > min_count).all()
-            ):  # nb_neighbours is >= 1
-                nb_pixels += 1
-                if interpolate == "interp_isolated":
-                    data[pixrow, pixcol] = temp.mean()
-                    mask[pixrow, pixcol] = 0
-                else:
-                    mask[pixrow, pixcol] = 1
-    else:  # 3D
-        for indx in range(target_pixels.shape[0]):
-            pix_z = target_pixels[indx, 0]
-            pix_y = target_pixels[indx, 1]
-            pix_x = target_pixels[indx, 2]
-            temp = data[
-                pix_z - extent : pix_z + extent + 1,
-                pix_y - extent : pix_y + extent + 1,
-                pix_x - extent : pix_x + extent + 1,
-            ]
-            temp = temp[np.logical_and(~np.isnan(temp), temp != target_val)]
-            if (
-                temp.size >= nb_neighbours and (temp > min_count).all()
-            ):  # nb_neighbours is >= 1
-                nb_pixels += 1
-                if interpolate == "interp_isolated":
-                    data[pix_z, pix_y, pix_x] = temp.mean()
-                    mask[pix_z, pix_y, pix_x] = 0
-                else:
-                    mask[pix_z, pix_y, pix_x] = 1
-
-    if debugging:
-        gu.combined_plots(
-            tuple_array=(data, mask),
-            tuple_sum_frames=(True, True),
-            tuple_sum_axis=(0, 0),
-            tuple_width_v=(None, None),
-            tuple_width_h=(None, None),
-            tuple_colorbar=(True, True),
-            tuple_vmin=(-1, 0),
-            tuple_vmax=(np.nan, 1),
-            tuple_scale=("log", "linear"),
-            tuple_title=("Data after filtering", "Mask after filtering"),
-            reciprocal_space=True,
-        )
-
-    return data, nb_pixels, mask
-
-
 def reload_bcdi_data(
     data,
     mask,
@@ -1887,7 +1438,7 @@ def reload_bcdi_data(
         )
 
         print("Intensity normalization using " + normalize_method)
-        data, monitor = util.normalize_dataset(
+        data, monitor = diff.normalize_dataset(
             array=data,
             monitor=monitor,
             norm_to_min=True,
